Немного изменил сущности
Поменял структуру Domain. Поменял названия некоторых сущностей. Добавил атрибуты Owned и навигационные свойства.
