--- Domain/AgeRating.cs
@@ -1,8 +1,0 @@
-﻿namespace Domain;
-
-// owned
-public class AgeRating
-{
-    public int? Age { get; set; }
-    public string? AgeMpaa { get; set; }
-}
