--- Domain/CurrencyValue.cs
@@ -1,8 +1,0 @@
-﻿namespace Domain;
-
-// Owned
-public class CurrencyValue
-{
-    public int Value { get; set; }
-    public string Currency { get; set; }  = null!;
-}
