+++ Domain/Entities/Budget.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace Domain;
+
+[Owned]
+public class Budget
+{
+    public int BudgetValue { get; set; }
+    public string BudgetCurrencyName { get; set; }  = null!;
+}
