+++ Domain/Entities/ContentType.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain;
+
+public class ContentType
+{
+    public int Id { get; set; }
+    public string ContentTypeName { get; set; } = null!;
+}
