+++ Domain/Entities/FavouriteContent.cs
