+++ Domain/Entities/PersonInContent.cs
@@ -3,7 +3,12 @@
 public class PersonInContent
 {
     public int Id { get; set; }
-    public int ContentId { get; set; }
+
+	public int ContentId { get; set; }
+	public ContentBase Content { get; set; }
+
     public string Name { get; set; } = null!;
+
+    public int ProfessionId { get; set; }
     public Profession Profession { get; set; } = null!;
 }
