+++ Domain/Entities/SeasonInfo.cs
