+++ Domain/Entities/SerialContent.cs
