+++ Domain/Entities/Subscription.cs
@@ -4,7 +4,7 @@ public class Subscription
 {
     public int Id { get; set; }
     public string Name { get; set; } = null!;
-    public DateTime ExpiresAt { get; set; }
-    public DateTime BoughtAt { get; set; }
+    public DateTimeOffset ExpiresAt { get; set; }
+    public DateTimeOffset BoughtAt { get; set; }
     public List<ContentBase> AccessibleContent { get; set; } = null!;
 }
