+++ Domain/Entities/TrailerInfo.cs
@@ -1,6 +1,8 @@
-﻿namespace Domain;
+﻿using Microsoft.EntityFrameworkCore;
 
-//Owned
+namespace Domain;
+
+[Owned]
 public class TrailerInfo
 {
     public string Url { get; set; } = null!;
