+++ Domain/Entities/YearRange.cs
@@ -1,6 +1,8 @@
-﻿namespace Domain;
+﻿using Microsoft.EntityFrameworkCore;
 
-//Owned
+namespace Domain;
+
+[Owned]
 public class YearRange
 {
     public DateOnly Start { get; set; }
