--- Domain/Type.cs
@@ -1,7 +1,0 @@
-﻿namespace Domain;
-
-public class Type
-{
-    public int Id { get; set; }
-    public string ContentType { get; set; } = null!;
-}
